requirements.txt
# MemoryOS 依赖
# 安装：pip install -r requirements.txt

# AI API 客户端
anthropic>=0.97.0
openai>=2.32.0

# 文件提取
pypdf>=6.0.0
python-docx>=1.2.0
openpyxl>=3.1.0

# 向量存储 & 检索
chromadb>=1.5.0
scikit-learn>=1.8.0
numpy>=2.0.0
rank-bm25>=0.2.2
jieba>=0.42.1
tiktoken>=0.12.0

# 后端服务
fastapi>=0.136.0
uvicorn>=0.46.0
aiohttp>=3.13.0
brotli>=1.2.0

# MCP 协议
mcp[cli]>=1.27.0

# 工具
python-dotenv>=1.2.0
rich>=15.0.0
tqdm>=4.67.0
pydantic>=2.13.0
